+++ pkg/query/macros_test.go
@@ -47,6 +47,37 @@ func TestInterPolateCombineValueMacros(t *testing.T) {
 		})
 	}
 }
+func TestInterPolateFromToMacros(t *testing.T) {
+	tests := []struct {
+		name          string
+		query         string
+		pluginContext backend.PluginContext
+		want          string
+		wantError     error
+	}{
+		{query: "foo", want: "foo"},
+		{query: "${__from}", want: "1500376552001"},
+		{query: "${__from:date}", want: "2017-07-18T12:15:52.001Z"},
+		{query: "${__from:date:iso}", want: "2017-07-18T12:15:52.001Z"},
+		{query: "foo ${__from:date:YYYY:MM:DD:hh:mm} bar", want: "foo 2017:07:18:12:15 bar"},
+		{query: "foo ${__to:date:YYYY-MM-DD:hh,mm} bar", want: "foo 2017-07-18:12,15 bar"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := querySrv.InterPolateMacros(tt.query, backend.TimeRange{
+				From: time.Unix(0, 1500376552001*1e6),
+				To:   time.Unix(0, 1500376552002*1e6),
+			}, tt.pluginContext)
+			if tt.wantError != nil {
+				require.NotNil(t, err)
+				require.Equal(t, tt.wantError, err)
+				return
+			}
+			require.Nil(t, err)
+			require.Equal(t, tt.want, got)
+		})
+	}
+}
 func TestInterPolateCustomIntervalMacros(t *testing.T) {
 	tests := []struct {
 		name          string
@@ -89,7 +120,6 @@ func TestInterPolateCustomIntervalMacros(t *testing.T) {
 		})
 	}
 }
-
 func TestApplyMacros(t *testing.T) {
 	tests := []struct {
 		name          string